Provide and recommend HTTP middleware
Pivot the package to not only provide an interceptor, but also HTTP
middleware. Both take the same authentication function, so it's easy to
swap between them. For efficiency and safety, documentation recommends
the HTTP middleware rather than the interceptor.

This also moves the module to use my new vanity domain.